Updated original metadata-list limit test
